Finally got to pixel data

+++ vme/DicomDictionary.cpp
@@ -12,9 +12,9 @@ CDicomDictionary::~CDicomDictionary(void)
 }
 
 
-map<unsigned short, string> CDicomDictionary::InitMap()
+map<unsigned int, string> CDicomDictionary::InitMap()
 {
-	map<unsigned short, string> mp;
+	map<unsigned int, string> mp;
 	mp[0x00020000] = "ULFile Meta Information Group Length";
 	mp[0x00020001] = "OBFile Meta Information Version";
 	mp[0x00020002] = "UIMedia Storage SOP Class UID";
